requirements.txt
Django
django-allauth
django-allauth-ui
psycopg2-binary
django-environ